auto-generate io_utils
