+++ src/options/io_utils_template.h
@@ -21,6 +21,7 @@
 #include <iosfwd>
 
 #include "options/language.h"
+#include "options/printer_options.h"
 
 /**
  * A collection of utilities to apply options that change how we print objects
@@ -37,31 +38,10 @@
  * `get*()` retrieves the specified option.
  */
 namespace cvc5::internal::options::ioutils {
-/** Set the default dag threshold */
-void setDefaultDagThresh(int64_t value);
-/** Set the default node depth */
-void setDefaultNodeDepth(int64_t value);
-/** Set the default output language */
-void setDefaultOutputLang(Language value);
 
-/** Apply the given dag threshold to the ios object */
-void applyDagThresh(std::ios_base& ios, int64_t dagThresh);
-/** Apply the given node depth to the ios object */
-void applyNodeDepth(std::ios_base& ios, int64_t nodeDepth);
-/** Apply the given output language to the ios object */
-void applyOutputLang(std::ios_base& ios, Language outputLang);
-/** Apply the given values to the ios object */
-void apply(std::ios_base& ios,
-           int64_t dagThresh,
-           int64_t nodeDepth,
-           Language outputLang);
-
-/** Get the dag threshold from the ios object */
-int64_t getDagThresh(std::ios_base& ios);
-/** Get the node depth from the ios object */
-int64_t getNodeDepth(std::ios_base& ios);
-/** Get the output language from the ios object */
-Language getOutputLang(std::ios_base& ios);
+// clang-format off
+${iodecls}$
+// clang-format on
 
 /**
  * A scope to copy and restore the options on an `std::ios_base` object in an
@@ -80,12 +60,10 @@ class Scope
  private:
   /** The ios object */
   std::ios_base& d_ios;
-  /** The stored dag threshold */
-  int64_t d_dagThresh;
-  /** The stored node depth */
-  int64_t d_nodeDepth;
-  /** The stored output language */
-  Language d_outputLang;
+
+// clang-format off
+${ioscope_members}$
+// clang-format on
 };
 }  // namespace cvc5::internal::options::ioutils
